Fix reversing the current order in Plone 4.
When the getObjPositionInParent script is not available in a folder
(this is the case in Plone 4), we do not show the option to sort on
position.

plone.folder 1.0.5 (which was included first in Plone 4.3.3) improves
things.  It makes it possible to simply reverse the current sort
order, which is what we want to offer.  So if this or a later version
is available, we offer simple reversing.

We make the 'id' keyword argument optional in the folder_position
script.  This is needed for the simple reversing.